pyproject.toml
[project]
name = "minesweeper"
version = "0.1.0"
description = "A minimal HUD environment"
requires-python = ">=3.11"
dependencies = [
 "anthropic>=0.64.0",
 "datasets>=4.0.0",
 "hud-python",
 "openai>=1.102.0",
]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[tool.hud]
image = "minesweeper:dev"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = [ "src/hud_controller",]
